Minor fixes; still need to resolve shared base_schema definitions; all based on Issue #54
